Encapsulates human behavior to facilitate the addition of other actions, such as travel to learn, etc.
