+++ simulation.py
@@ -9,10 +9,7 @@ from config import Configuration, config_error
 from environment import build_hospital
 from infection import find_nearby, infect, recover_or_die, compute_mortality,\
 healthcare_infection_correction
-from motion import update_positions, out_of_bounds, update_randoms,\
-get_motion_parameters
-from path_planning import go_to_location, set_destination, check_at_destination,\
-keep_at_destination, reset_destinations
+from motion import Motion
 from population import initialize_population, initialize_destination_matrix,\
 set_destination_bounds, save_data, save_population, Population_trackers
 from visualiser import build_fig, draw_tstep, set_style, plot_sir
@@ -37,6 +34,9 @@ class Simulation():
 
         #initalise destinations vector
         self.destinations = initialize_destination_matrix(self.Config.pop_size, 1)
+
+        #Control all human behavior here
+        self.motion_control = Motion()
 
 
     def reinitialise(self):
@@ -64,54 +64,11 @@ class Simulation():
             #initialize figure
             self.fig, self.spec, self.ax1, self.ax2 = build_fig(self.Config)
 
-        #check destinations if active
-        #define motion vectors if destinations active and not everybody is at destination
-        active_dests = len(self.population[self.population[:,11] != 0]) # look op this only once
+        #Monitor the basic actions of all people such as moving, out of bounding, or dead.
+        self.motion_control.set_general_rule_motion(self.population,self.destinations,self.Config)
 
-        if active_dests > 0 and len(self.population[self.population[:,12] == 0]) > 0:
-            self.population = set_destination(self.population, self.destinations)
-            self.population = check_at_destination(self.population, self.destinations,
-                                                   wander_factor = self.Config.wander_factor_dest,
-                                                   speed = self.Config.speed)
-
-        if active_dests > 0 and len(self.population[self.population[:,12] == 1]) > 0:
-            #keep them at destination
-            self.population = keep_at_destination(self.population, self.destinations,
-                                                  self.Config.wander_factor)
-
-        #out of bounds
-        #define bounds arrays, excluding those who are marked as having a custom destination
-        if len(self.population[:,11] == 0) > 0:
-            _xbounds = np.array([[self.Config.xbounds[0] + 0.02, self.Config.xbounds[1] - 0.02]] * len(self.population[self.population[:,11] == 0]))
-            _ybounds = np.array([[self.Config.ybounds[0] + 0.02, self.Config.ybounds[1] - 0.02]] * len(self.population[self.population[:,11] == 0]))
-            self.population[self.population[:,11] == 0] = out_of_bounds(self.population[self.population[:,11] == 0],
-                                                                        _xbounds, _ybounds)
-
-        #set randoms
-        if self.Config.lockdown:
-            if len(self.pop_tracker.infectious) == 0:
-                mx = 0
-            else:
-                mx = np.max(self.pop_tracker.infectious)
-
-            if len(self.population[self.population[:,6] == 1]) >= len(self.population) * self.Config.lockdown_percentage or\
-               mx >= (len(self.population) * self.Config.lockdown_percentage):
-                #reduce speed of all members of society
-                self.population[:,5] = np.clip(self.population[:,5], a_min = None, a_max = 0.001)
-                #set speeds of complying people to 0
-                self.population[:,5][self.Config.lockdown_vector == 0] = 0
-            else:
-                #update randoms
-                self.population = update_randoms(self.population, self.Config.pop_size, self.Config.speed)
-        else:
-            #update randoms
-            self.population = update_randoms(self.population, self.Config.pop_size, self.Config.speed)
-
-        #for dead ones: set speed and heading to 0
-        self.population[:,3:5][self.population[:,6] == 3] = 0
-
-        #update positions
-        self.population = update_positions(self.population)
+        #special event happen(such as lockdown) and set randoms
+        self.motion_control.special_event(self.population,self.destinations,self.Config,self.pop_tracker)
 
         #find new infections
         self.population, self.destinations = infect(self.population, self.Config, self.frame,
